Minor changes
Former-commit-id: 7c2a6cef2eafecc87f99b159db9fd9d1e36e7672

+++ src/def.hpp
@@ -30,7 +30,7 @@ using u32    = unsigned int;
 using i64    = signed long long;
 using u64    = unsigned long long;
 using dur_t  = std::chrono::duration<double>;
-using prec_t = float;  // Precision type
+using prec_t = double;  // Precision type -- MUST be double
 
 // Metaprogram
 // Power a^n
